@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
@@ -7,13 +7,23 @@ import { CookiesProvider } from "react-cookie";
 import { Provider } from "react-redux";
 import { store } from "./store"
 
-ReactDOM.render(
+// ReactDOM.render(
+//   <Provider store={store}>
+//     <CookiesProvider>
+//       <App />
+//     </CookiesProvider>
+//   </Provider>,
+//   document.getElementById("root")
+// );
+
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
   <Provider store={store}>
     <CookiesProvider>
       <App />
     </CookiesProvider>
-  </Provider>,
-  document.getElementById("root")
+  </Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
